+++ get_next_line.h
@@ -3,25 +3,26 @@
 /*                                                        :::      ::::::::   */
 /*   get_next_line.h                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fschuber <fschuber@student.42.fr>          +#+  +:+       +#+        */
+/*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/19 10:25:52 by fschuber          #+#    #+#             */
-/*   Updated: 2023/10/23 07:39:36 by fschuber         ###   ########.fr       */
+/*   Updated: 2023/10/23 15:55:20 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-
-#ifndef BUFFER_SIZE
-# define BUFFER_SIZE 42
-#endif
 
 #ifndef GET_NEXT_LINE_H
 # define GET_NEXT_LINE_H
 
-#include <unistd.h>
-#include <stdlib.h>
+# include <unistd.h>
+# include <stdlib.h>
 
-int		ft_strlen(char	*str);
-char	*gnl_substr(char const *s, unsigned int start, size_t len);
+# ifndef BUFFER_SIZE
+#  define BUFFER_SIZE 42
+# endif
+
+char	*get_next_line(int fd);
+int		ft_strlen(const char	*str);
+char	*gnl_substr(char const *s, int start, size_t len);
 char	*ft_strjoin(char const *s1, char const *s2);
 
 #endif