+++ changelog.md
@@ -1,5 +1,19 @@
 # Commit Changelog get_next_line by @fschuber
 
+#### Potential Future Improvements
+- Change strjoin to use realloc instead of new malloc
+- on the first iteration when leftovers is null, weird shit will happen with strjoin. instead, change strjoin to return the other string if either string is null
+
+#### v0.3.1
+- strjoin while condition logic + typo fix
+- read return value fix
+- Norminette Improvements
+- General Improvements
+- Corrected Header Order
+- Memory leaks
+	- buffer now gets freed when read fails
+	- previous leftovers content gets freed on strjoin
+
 ### v0.3.0
 - third iteration, recursive approach
 - not tested